Change inventory search to be based on location
This is a bit of a hack. The location is hardcoded to Monmouth. The
point is to reduce the inventory volume to stores (near Monmouth) that I
might actually care about. The reduced data means we can avoid monkeying
with the page size for a while. All this will be revisited.