Add description to OlccBottle
The attribute is actually called "description" on the OLCC website - it
should have been called this all along. We have a migration to copy the
name attribute to description for all existing bottles, and when we
create new bottles in the UpdateCategoryBottles service, we'll copy it
over there, too. This will allow us to (manually) edit the "name"
attribute to create more user-friendly names in the UI without having to
modify any of the views.

A more complete/better way to handle this would be to rename the
attribute to "description" in the DTO (and the HTML parsing logic), and
add some before_save hook to the model to copy from description to name
if name is empty/nil.