Improve handling of bottle name vs. description
Quit being lazy and fixed the DTO to use description. That actually
caused an issue in production where all the fixed-up names got reset and
events were generated for all of them.
There is still an edge case to deal with: when descriptions are
legitimately updated (at OLCC). What we need to update the name unless
it was already manually edited (maybe).