Add error messages to ApplicationService
The LoadBottle service is the first one to use it, but only for OLCC
errors. Although the bottles controller could use the error message in
the flash, I'm not going to bother right now out of laziness plus that
function to add bottles is a bit of a (temporary?) hack.
Part of #116