Add Devise for real signup and login
This is WIP for #122 - the first, big step

* Initial work to add Devise to project
As a result, users are always not logged in. No views for setting up
users or logging in. Passwords are now required for users - fixed the
factory for that.
* Add generated Devise views
Added some links on home/index to access login/logout.
I'm committing the Mailer configs, but I haven't tested anything yet.
* Set up config for Gmail as SMTP provider
Moved user and password into environment variables for development. We
will eventually need the same config for production when this is ready.