123 - Add user's categories to recent event filter
The list of categories is still hard-coded on User but new bottle events
will only show up for bottles in the categories that the user follows.
Turns out, when you combine "joins" with "includes", ActiveRecord does
not need to run a second query to pick up the N included objects - so
smart.